No break build userspace. 5.10 kernel now broken, revert to 5.9
Revert "Merge branch 'master' into imx6-ravion-vanilla"

This reverts commit dad864d5968ea8f7565f19f9003d828c22556c18, reversing
changes made to 063cd4ff05f60244bafe0332110e8536221a6a69.

+++ include/media/videobuf-dma-sg.h
@@ -60,7 +60,7 @@ struct videobuf_dmabuf {
 	/* common */
 	struct scatterlist  *sglist;
 	int                 sglen;
-	unsigned long       nr_pages;
+	int                 nr_pages;
 	int                 direction;
 };
